requirements.txt
pyyaml
opencv-python-headless
numpy
pytesseract
pdf2image
moviepy
scikit-image
tqdm
matplotlib
fuzzywuzzy
python-Levenshtein
easyocr
openai
transformers
pydantic
pytz
torchaudio
rich
PyPDF2 
reportlab